Change dhcp template to use new format
We now have 2 Metalware commands for working with dhcpd templates,
`hunter` and `dhcp`. `hunter` builds up a config file of hostnames to
MAC addresses, while `dhcp` uses this config file to write an entire
`dhcp` template.

This commit adapts the existing `dhcp` template to produce a `dhcp`
config for all nodes using the `hunter` config format, making it
suitable to use with the new `dhcp` command.