Pull IP/hostnames from node network config
Rather than defining hostname in Bash various places, can now
dynamically define this within the network config for each node, and
then use these values in templates where needed.

We were also looking up the IP for nodes based on the hostname, but we
can define this in the config too; I'm leaving the lookup functionality
for now in case an IP is unset, but may we want to remove this later.