Correct dhcp template to access domain directly
Previously we defined the full hostname for a hunted node on the primary
network as `<%= hostname %><%= networks.pri.hostname %>`, which resolves
down to `<%= hostname %><%= alces.nodename %>.pri.<%= domain %>`.

However, when this template is rendered we are not in a 'node' context,
and so `alces.nodename` is `nil`; instead we are iterating through each
of the hunted nodes in a 'domain' context, with only the domain configs
and answers loaded.

Since `alces.nodename` is `nil`, the above does resolve to the correct
value, however it is slightly confusing to rely on a node-related value
with part of it unset when we are not templating for a node. Instead we
now form the hostname using the `domain` value directly, making what is
going on more obvious.