Add example of node using basic build_method
This commit adds a simple default `basic` `build_method` template,
similar to the 'post' section in the current Kickstart template.

When run, the rendered script from this template should simply request
the same files as would be requested when Kickstart-building a node, and
then report the node as built. Also included is a simple config for a
node `self` which would be built using this method; we may or may not
want to keep this config long term.

This uses the new `basic` `build_method` feature added in
https://github.com/alces-software/metalware/pull/156.