Remove localhost stuff from hosts template
We now only render part of the hosts file with Metalware, and leave
existing content outside this section unchanged (see
https://github.com/alces-software/metalware/pull/158).

Therefore, we don't need to include these lines for managing localhost
or the Metalware deployment server; if the user wants to include these
lines they can now simply leave them in when setting up their server.

Including the deployment server line in the template itself could also
have been error prone, as if this line broke then future renders would
also not work as we determine the `hostip` using this, so having the
line defining this IP managed outside Metalware is also a good idea.