Give genders_additional_groups explicit default
This seems better than using `optional`, as it amounts to the same thing
but makes explicit what the value of this question will be in templates
by default (previously it would have a value of `nil`, which would just
become an empty string when inserted into a template).

We now have no values specified as `optional` in `configure.yaml` -
possibly we can remove support for this entirely from Metalware for the
same reason?

This change was prompted by it being tricky to handle `optional` values
in the Metalware GUI, while changing this to use `default` means no
special handling is needed.