Add initial running of plugin scripts
This gives an example of how scripts specified as plugin files can be
run along with the other repo files.

Note that the accessing of plugin file sections like
`(plugin.files.setup || [])` is necessary as not every plugin may
contain every possible files section, and if a files section does not
exist then we get `nil` instead. To ensure we do not error in this
situation we must use an empty array rather than attempt to call `each`
on the `nil` section.